services/requirements.dev.txt
# Core Django & REST
Django>=5.0,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0

social-auth-app-django>=5.4,<5.6

# Environment & Config
python-dotenv>=1.0.1

# Google Cloud Secret Manager (for hybrid config)
#google-cloud-secret-manager>=2.20.0

# (Optional) Database
#psycopg2-binary>=2.9.9  # if using PostgreSQL

# Celery
celery~=5.3.6

# Developer tools
black>=24.4.2
flake8>=7.0.0

# Testing
pytest>=8.2.0
pytest-django>=4.8.0
